ECR Viwer Demo Site (#31)
* ecr veiwer

* fix ecrtable test

* Remove margin of ecrtable

* Move font css to body

* Fix button links

* missed router

* Update router

* limit ECRTable size

* rebase

* fix ecrtable test

* rebase

* Move font css to body

* Fix button links

* missed router

* Update router

* limit ECRTable size

---------
