+++ front-end/components/ECRTable/ECRTable.css
@@ -1,0 +1,3 @@
+.usa-table {
+    max-width: 512px;
+}
