+++ front-end/tests/components/ECRTable.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import {render, screen} from '@testing-library/react';
 import "@testing-library/jest-dom";
-import ECRTable from '../../components/ECRTable'; // Adjust the import path as per your project structure
+import ECRTable from '../../components/ECRTable/ECRTable'; // Adjust the import path as per your project structure
 
 describe('ECRTable', () => {
     it('renders ECRTable component with provided data', () => {
@@ -19,11 +19,10 @@ describe('ECRTable', () => {
             }
         };
 
-        render(<ECRTable ecrData={ecrData} />);
+        render(<ECRTable ecrData={ecrData}/>);
 
         // You can add more specific assertions here
         // For example, checking if the table headers and data are rendered
-        expect(screen.getByText('eCR Viewer')).toBeInTheDocument();
         expect(screen.getByText('Field Name')).toBeInTheDocument();
         expect(screen.getByText('Field Value')).toBeInTheDocument();
         expect(screen.getByText('Patient Id')).toBeInTheDocument();
